+++ R/call.R
@@ -900,6 +900,48 @@ is_call_simple <- function(x, ns = NULL) {
   namespaced || is_symbol(head)
 }
 
+is_call_index <- function(x, ns = NULL) {
+  check_required(x)
+
+  if (!is_call(x)) {
+    return(FALSE)
+  }
+
+  out <- FALSE
+
+  while (is_call(fn <- x[[1]])) {
+    if (!is_call(fn, c("$", "@", "[", "[["))) {
+      return(FALSE)
+    }
+
+    if (!every(fn[-1], is_arg_index, ns)) {
+      return(FALSE)
+    }
+
+    out <- TRUE
+    x <- fn
+  }
+
+  out
+}
+
+is_arg_index <- function(arg, ns) {
+  if (!is_call(arg)) {
+    return(TRUE)
+  }
+
+  namespaced <- is_call(arg, c("::", ":::"))
+  if (namespaced) {
+    if (!is_null(ns) && !identical(namespaced, ns)) {
+      return(FALSE)
+    } else {
+      return(TRUE)
+    }
+  }
+
+  is_call_simple(arg)
+}
+
 #' Extract arguments from a call
 #'
 #' @inheritParams call_name